ap - combine admin and regular user tests into one file
